# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MTB REST API
  version: 1.0.0
  description: REST API of the MTB service.
...
servers:
  - url: /api
paths:
...
  /episodes:
    #    get:
    #      summary: List all available episodes.
    #      operationId: getEpisodes
    #      tags:
    #        - episode
    #      responses:
    #        200:
    #          description: OK
    #          content:
    #            application/json:
    #              schema:
    #                type: array
    #                items:
    #                  $ref: '#/components/schemas/Episode'
    #        401:
    #          description: Unauthorized - The provided credentials are unknown
    #    delete:
    #      summary: delete all available episodes.
    #      operationId: deleteEpisodes
    #      tags:
    #        - episode
    #      responses:
    #        200:
    #          description: OK
    #        401:
    #          description: Unauthorized - The provided credentials are unknown
    post:
      summary: Add new episode to database.
      description: Add new episode to database.
      operationId: addEpisode
      tags:
        - episode
      requestBody:
        description: Create a new episode in database.
        content:
          application/json:
            schema:
              type: object
              properties:
                patientId:
                  type: string
                  format: uuid
                workflowId:
                  type: integer
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        400:
          description: Foreign key constraint violated (Patient or Workflow not found)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        405:
          description: Invalid Input
  /episodes/{id}:
    get:
      summary: Get one Episode.
      operationId: getEpisode
      tags:
        - episode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Episode not found
    put:
      summary: Update one Episode without updating stepInfo.
      operationId: updateEpisode
      tags:
        - episode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      requestBody:
        description: The updated episode
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Episode'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        400:
          description: Foreign key constraint violated (Patient or Workflow is not the same as in the saved episode)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Episode not found
...
  /labnumbers:
    post:
      summary: Add new LabNumber Entry to the database.
      description: Add new lab number.
      operationId: addLabNumber
      tags:
        - lab_number
      requestBody:
        description: Create a new Entry of labNumber database.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabNumber'
        required: true
      responses:
        200:
          description: OK
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: Patient not found
        405:
          description: Invalid Input (LabNumberId ist already assigned to another patient)
    get:
      summary: Get all LabNumber entries
      operationId: getLabNumbers
      tags:
        - lab_number
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabNumber'
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
  /labnumbers/{labNumberId}:
    get:
      summary: Get LabNumber Entry.
      operationId: getLabNumber
      tags:
        - lab_number
      parameters:
        - name: labNumberId
          in: path
          required: true
          schema:
            type: string
          description: The labnumber indentifer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabNumber'
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: LabNumber not found
  /labnumbers/{labNumberId}/checkLabNumber:
    get:
      summary: Checks if LabNumber is already assigned.
      operationId: isLabNumberExists
      tags:
        - lab_number
      parameters:
        - name: labNumberId
          in: path
          required: true
          schema:
            type: string
          description: The labnumber indentifer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: LabNumber not found
...
components:
  schemas:
    LabNumber:
      type: object
      properties:
        id:
          type: string
        specimenLabelling:
          type: string
        episodeId:
          type: string
          format: uuid
        assigned:
          type: boolean
        assignedOn:
          type: string
          format: date-time
          example: '2020-05-27T17:32:28Z'
        createdAt:
          type: string
          format: date-time
          example: '2020-05-27T17:32:28Z'
...
    Episode:
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        requirementId:
          type: string
          format: uuid
        report:
          type: string
        decision:
          type: string
        stepInfo:
          type: array
          items:
            $ref: '#/components/schemas/StepInfo'
        createdAt:
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
        updatedAt:
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
        workflowId:
          type: string
          format: integer
      required:
        - patientId
        - workflowId
...
    StepInfo:
      type: object
      properties:
        stepStatus:
          $ref: '#/components/schemas/StepStatus'
        stepId:
          type: integer
        episodeId:
          type: string
          format: uuid
    StepStatus:
      type: string
      enum: [ PENDING, INCOMPLETE, COMPLETE ]
...